# Fixed checkodesol function