--- a/sympy/solvers/ode.py
+++ b/sympy/solvers/ode.py
@@ -338,13 +338,10 @@ lie_heuristics = (
     )
 
 
-def sub_func_doit(eq, func, new, order=None):
+def sub_func_doit(eq, func, new):
     r"""
     When replacing the func with something else, we usually want the
     derivative evaluated, so this function helps in making that happen.
-    If ``order`` is provided, then all derivatives of all orders will
-    be evaluated; otherwise, only the derivatives present (default)
-    will be evaluated.
 
     Examples
     ========
@@ -362,15 +359,14 @@ def sub_func_doit(eq, func, new, order=None):
     x*(-1/(x**2*(z + 1/x)) + 1/(x**3*(z + 1/x)**2)) + 1/(x*(z + 1/x))
     ...- 1/(x**2*(z + 1/x)**2)
     """
-    reps = {func: new}
-    if order:
-        x = func.args[0]
-        for i in range(order, -1, -1):
-            reps[func.diff(x, i)] = new.diff(x, i)
-    else:
-        for d in eq.atoms(Derivative):
-            reps[d] = d.subs(func, new).doit(deep=False)
+    reps= {func: new}
+    for d in eq.atoms(Derivative):
+        if d.expr == func:
+            reps[d] = new.diff(*d.variable_count)
+        else:
+            reps[d] = d.xreplace({func: new}).doit(deep=False)
     return eq.xreplace(reps)
+
 
 def get_numbered_constants(eq, num=1, start=1, prefix='C'):
     """
@@ -2404,7 +2400,7 @@ def checkodesol(ode, sol, func=None, order='auto', solve_for_func=True):
             ode_diff = ode.lhs - ode.rhs
 
             if sol.lhs == func:
-                s = sub_func_doit(ode_diff, func, sol.rhs, order)
+                s = sub_func_doit(ode_diff, func, sol.rhs)
             else:
                 testnum += 1
                 continue