@jmig5776 2019-02-10
@smichr Please review it as you are familiar with `checkodesol`.