@oscarbenjamin 2019-02-10
Is it necessary to call `doit`? That can have other effects since `doit` forces evaluation of many slow things. I don't know whether it would generally speed up or slow down `checkodesol` though. In #15913 I gave a solution that didn't involve calling `doit` and was much faster than `checkodesol`:
```python
simplify(neq.subs(f(x).diff(x), wsol.diff(x)).subs(f(x), wsol))
```
Is there some reason that `checkodesol` doesn't just do that in general? That seems like the obvious way to implement the function to me...

We should check the time taken to run the `ode` test suite before and after this change. @jmig5776  do you know how to run the tests from `test_ode.py` yourself on your computer?