@oscarbenjamin 2019-02-10
I compared this PR with master under 3.7 when running the not slow tests with
```console
$ bin/test sympy/solvers/tests/test_ode.py 
```
This gives 307 seconds for master and 355 seconds for this PR which is a significant slowdown.

We should check also what happens with the slow tests.