@oscarbenjamin 2019-02-10
Running slow tests as well (but not the tests that are both slow and xfail - some of those should never be run) using pytest I get:
```terminal
$ pytest sympy/solvers/tests/test_ode.py -m 'not slow or not xfail' -n4
====================================================== test session starts =======================================================
platform darwin -- Python 3.7.1, pytest-4.1.2.dev19+g57bf9d67, py-1.7.0, pluggy-0.8.0
architecture: 64-bit
cache:        yes
ground types: gmpy 1.17

rootdir: /Users/enojb/current/sympy/sympy, inifile:
plugins: xdist-1.26.0, forked-0.2, cov-2.6.1, doctestplus-0.3.0.dev0
gw0 [110] / gw1 [110] / gw2 [110] / gw3 [110]
...........x............s...................xxx................................x..................x.xx.xx.x...             [100%]
======================================= 98 passed, 1 skipped, 11 xfailed in 318.26 seconds =======================================
```
for this PR and
```terminal
$ pytest sympy/solvers/tests/test_ode.py -m 'not slow or not xfail' -n4
====================================================== test session starts =======================================================
platform darwin -- Python 3.7.1, pytest-4.1.2.dev19+g57bf9d67, py-1.7.0, pluggy-0.8.0
architecture: 64-bit
cache:        yes
ground types: gmpy 1.17

rootdir: /Users/enojb/current/sympy/sympy, inifile: pytest.ini
plugins: xdist-1.26.0, forked-0.2, cov-2.6.1, doctestplus-0.3.0.dev0
gw0 [111] / gw1 [111] / gw2 [111] / gw3 [111]
...........x............s....................x..........xx.................x.....x.................x.x.x..xx...            [100%]
======================================= 99 passed, 1 skipped, 11 xfailed in 260.38 seconds =======================================
```
under master.

Again that's a significant slowdown (318 seconds this PR and 260 seconds for master).

Trying this manually in ipython I can see that the simplify/subs method is much faster than checkodesol.
```julia
In [1]: C1 = Symbol('C1')                                                                                                         

In [2]: C2 = Symbol('C2')                                                                                                         

In [3]: neq = -C1/x - 2*x*f(x) - f(x) + Derivative(f(x), x)                                                                       

In [4]: wsol = C2*exp(x**2 + x) + exp(x**2 + x)*Integral(C1*exp(-x**2 - x)/x, x)                                                  

In [5]: simplify(neq.subs(f(x).diff(x), wsol.diff(x)).subs(f(x), wsol))   # returns instantly                                                         
Out[5]: 0

In [6]: checkodesol(neq, wsol)    # takes a few seconds                                                                                                
Out[6]: (True, 0)
```