@jmig5776 2019-02-10
>Is there some reason that `checkodesol` doesn't just do that in general? That seems like the obvious way to implement the function to me...

Actually here the sub_func_doit function returns a different form of lhs when i=1 so thats why i used doit in last.
https://github.com/sympy/sympy/blob/a7aed50f600c79c9b977b5e8059346312e625716/sympy/solvers/ode.py#L2504-L2507