@oscarbenjamin 2019-02-10
What happens if you change the function so that before trying all the things it currently does it tries what I suggested above?