@jmig5776 2019-02-10
If i do like this i.e doit when f(x) is substituted
```
lhs = sub_func_doit(lhs, func.diff(x, i), diffsols[i])
                    rhs = sub_func_doit(rhs, func.diff(x, i), diffsols[i])
                    ode_or_bool = Eq(lhs, rhs)
                    if i==0:
                        ode_or_bool = ode_or_bool.doit()
                    ode_or_bool = simplify(ode_or_bool)
```
Then the tests are running like this
In master
```
sympy/solvers/tests/test_ode.py[111] ............f..w.w......w...w.......ww.f...
...w.wff.w.fww.......................w.....f.w..............f.f.ffff        [OK]

________________________________ slowest tests _________________________________
test_separable2 - Took 10.452 seconds
test_series - Took 11.287 seconds
test_linear_2eq_order2 - Took 13.494 seconds
test_separable_reduced - Took 19.812 seconds
test_2nd_power_series_ordinary - Took 24.201 seconds
 tests finished: 86 passed, 13 skipped, 12 expected to fail, in 217.20 seconds
```
In PR
```
sympy/solvers/tests/test_ode.py[112] ............f..w.w......w...w.......ww.f...
...w.wff.w.fww.......................w.....f.w..............f.f.ffff.       [OK]

________________________________ slowest tests _________________________________
test_separable2 - Took 10.517 seconds
test_series - Took 11.291 seconds
test_linear_2eq_order2 - Took 14.414 seconds
test_separable_reduced - Took 20.357 seconds
test_2nd_power_series_ordinary - Took 24.616 seconds
 tests finished: 87 passed, 13 skipped, 12 expected to fail, in 222.00 seconds 
```
What do you think about this idea?