@oscarbenjamin 2019-02-10
What difference dies this make to the timings? Does it seem faster when you use it manually?