@jmig5776 2019-02-10
>What difference dies this make to the timings? Does it seem faster when you use it manually?

For the given issue timing has been faster.