@oscarbenjamin 2019-02-10
> > What difference dies this make to the timings? Does it seem faster when you use it manually?
> 
> For the given issue timing has been faster.

That was not really a yes/no question. I was hoping for you to show a comparison with numbers (e.g. I ran this command and on master got this time and with this PR got another time. Also it is noticeably faster - or something like that).