@oscarbenjamin 2019-02-12
I just tried this
```diff
diff --git a/sympy/solvers/ode.py b/sympy/solvers/ode.py
index ad30197248..0ec8ba2359 100644
--- a/sympy/solvers/ode.py
+++ b/sympy/solvers/ode.py
@@ -2419,9 +2419,25 @@ def checkodesol(ode, sol, func=None, order='auto', solve_for_func=True):
             return checkodesol(ode, eqs, order=order,
                 solve_for_func=False)
 
-    s = True
+    # Used by the bizarre loop below...
     testnum = 0
     x = func.args[0]
+
+    # Straight-forward substitution of explicit solution
+    if sol.lhs == func and not sol.rhs.has(func):
+        solexpr = sol.rhs
+    elif sol.rhs == func and not sol.lhs.has(func):
+        solexpr = sol.lhs
+    else:
+        solexpr = None
+    if solexpr is not None:
+        ode_expr = ode.lhs - ode.rhs
+        for i in range(order,-1,-1):
+            ode_expr = ode_expr.subs(func.diff(x, i), solexpr.diff(x, i))
+        if simplify(ode_expr) == 0:
+            return (True, 0)
+
+    s = True
     while s:
         if testnum == 0:
             # First pass, try substituting a solved solution directly into the
```
and it made the test suite run faster.

On master not slow tests:
```console
$ bin/test sympy/solvers/tests/test_ode.py 
...
==== tests finished: 88 passed, 13 skipped, 12 expected to fail, in 295.92 seconds ====
```
On master including slow tests:
```console
$ pytest sympy/solvers/tests/test_ode.py -m 'not slow or not xfail' -n4
...
==== 99 passed, 1 skipped, 11 xfailed in 260.94 seconds ====
```
With diff not slow tests:
```console
$ bin/test sympy/solvers/tests/test_ode.py 
...
==== tests finished: 88 passed, 13 skipped, 12 expected to fail, in 287.72 seconds ====
```
With diff including slow tests:
```console
$ pytest sympy/solvers/tests/test_ode.py -m 'not slow or not xfail' -n4
...
==== 99 passed, 1 skipped, 11 xfailed in 251.93 seconds ====
```

Also the issue is solved with this:
```julia
In [7]: C1 = Symbol('C1')                                                                                                                                     

In [8]: C2 = Symbol('C2')                                                                                                                                     

In [9]: neq = -C1/x - 2*x*f(x) - f(x) + Derivative(f(x), x)                                                                                                   

In [10]: wsol = C2*exp(x**2 + x) + exp(x**2 + x)*Integral(C1*exp(-x**2 - x)/x, x)                                                                             

In [11]: checkodesol(neq, wsol)                                                                                                                               
Out[11]: (True, 0)
```