@oscarbenjamin 2019-02-13
My timings indicate that this PR shows a substantial slowdown for the ODE tests:

Master:
```console
$ bin/test sympy/solvers/tests/test_ode.py # 295 secs
$ pytest sympy/solvers/tests/test_ode.py -m 'not slow or not xfail' -n4 # 292 secs
```
This PR:
```console
$ bin/test sympy/solvers/tests/test_ode.py # 344 secs
$ pytest sympy/solvers/tests/test_ode.py -m 'not slow or not xfail' -n4 # 314 secs
```

The diff I showed above achieved (for me on this computer) better timings as well as solving the issue at hand.