@oscarbenjamin 2019-02-15
Okay so timings for non-slow tests are (using bin/test and test_ode.py from master in all cases):

Master: 305 seconds
This PR: 312 seconds
Master with this PR merged in: 292 seconds

So looks good on the timing front and I guess that's the best way to benchmark a PR in future.

I've run the slow tests with
```console
$ pytest sympy/solvers/tests/test_ode.py -m 'not slow or not xfail' -n4
```
and I get 283 seconds for this PR (merged into master) and 286 seconds for master.